app/requirements.txt
Flask
itsdangerous
Jinja2
MarkupSafe
psycopg2-binary
pymssql
Werkzeug